refactor: Swagger 인증 로직 및 FastAPI 문서화 관련 코드 정리
