+++ backend/app/main.py
@@ -1,29 +1,14 @@
-from fastapi import FastAPI, Depends, HTTPException, status, Response
-from fastapi.security import HTTPBasic, HTTPBasicCredentials
+from fastapi import FastAPI, Depends, status, Response
 from fastapi.openapi.docs import get_swagger_ui_html
 from fastapi.openapi.utils import get_openapi
-from typing import Annotated
 
 import os
 from contextlib import asynccontextmanager
 from core.db import db
 from api import api_router
 from starlette.middleware.cors import CORSMiddleware
+from core.dependencies import authenticate_user
 
-# Swagger 인증 정보
-SWAGGER_USERNAME = os.environ.get("SWAGGER_USERNAME", "admin")
-SWAGGER_PASSWORD = os.environ.get("SWAGGER_PASSWORD", "admin")
-
-# 인증 의존성
-security = HTTPBasic()
-
-def authenticate_user(credentials: Annotated[HTTPBasicCredentials, Depends(security)]):
-    if credentials.username != SWAGGER_USERNAME or credentials.password != SWAGGER_PASSWORD:
-        raise HTTPException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="Incorrect credentials",
-            headers={"WWW-Authenticate": "Basic"},
-        )
 
 # lifespan 유지
 @asynccontextmanager
@@ -42,11 +27,11 @@ app = FastAPI(
 
 # ✅ 실제 인증된 사용자에게만 문서 제공
 @app.get("/docs", include_in_schema=False)
-def get_docs(credentials: Annotated[HTTPBasicCredentials, Depends(authenticate_user)]):
+def get_docs(_=Depends(authenticate_user)):
     return get_swagger_ui_html(openapi_url="/openapi.json", title="Bingo API Docs")
 
 @app.get("/openapi.json", include_in_schema=False)
-def get_openapi_spec(credentials: Annotated[HTTPBasicCredentials, Depends(authenticate_user)]):
+def get_openapi_spec(_=Depends(authenticate_user)):
     return get_openapi(title="Bingo API", version="1.0.0", routes=app.routes)
 
 # 라우터 및 기타 설정
@@ -57,6 +42,7 @@ app.add_middleware(
     allow_credentials=True,
     allow_methods=["*"],
     allow_headers=["*"],
+    expose_headers=["Content-Disposition"],
 )
 
 
@@ -66,7 +52,7 @@ def hc():
 
 
 @app.post("/api/zozo-manual-reset-db")
-async def reset_db(response: Response):
+async def reset_db(response: Response, include_in_schema: bool = False):
     try:
         response.headers["Cache-Control"] = "no-store, no-cache, must-revalidate, proxy-revalidate"
         response.headers["Pragma"] = "no-cache"
